:zap: perf: Atualizando Migrations

+++ src/database/migrations/1727892942058-AddCarIdToCreateReserve.ts
@@ -1,19 +1,12 @@
-import {
-  MigrationInterface,
-  QueryRunner,
-  TableColumn,
-  TableForeignKey,
-} from 'typeorm';
+import { MigrationInterface, QueryRunner, TableColumn, TableForeignKey } from 'typeorm';
 
-export class AddCarIdToCreateReserve1727892942058
-  implements MigrationInterface
-{
+export class AddCarIdToCreateReserve1727892942058 implements MigrationInterface {
   public async up(queryRunner: QueryRunner): Promise<void> {
     await queryRunner.addColumn(
       'reserve',
       new TableColumn({
         name: 'car_id',
-        type: 'int',
+        type: 'integer',
         isNullable: true,
       }),
     );
@@ -21,17 +14,22 @@ export class AddCarIdToCreateReserve1727892942058
     await queryRunner.createForeignKey(
       'reserve',
       new TableForeignKey({
-        name: 'ReserveCar',
         columnNames: ['car_id'],
         referencedColumnNames: ['id'],
         referencedTableName: 'car',
-        onDelete: 'SET NULL',
+        onDelete: 'CASCADE',
       }),
     );
   }
 
   public async down(queryRunner: QueryRunner): Promise<void> {
-    await queryRunner.dropForeignKey('reserve', 'ReserveCar');
-    await queryRunner.dropColumn('reserve', 'car_id');
+    const table = await queryRunner.getTable('reserve');
+    if (table) {
+      const foreignKey = table.foreignKeys.find(fk => fk.columnNames.indexOf('car_id') !== -1);
+      if (foreignKey) {
+        await queryRunner.dropForeignKey('reserve', foreignKey);
+      }
+      await queryRunner.dropColumn('reserve', 'car_id');
+    }
   }
 }
